requirements.txt
geojson==3.1.0
gtfs_kit==6.1.0
pandas==2.2.2
requests==2.32.3
transitfeed-py3==1.2.16
geopandas==1.0.1
python-on-whales==0.72.0